REFACTOR - Application Nav Graph & Separate Files
