+++ app/src/main/java/com/varqulabs/diproveboliviapp/core/presentation/composables/DiproveCenterAppBar.kt
@@ -21,7 +21,6 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.varqulabs.diproveboliviapp.R
-import com.varqulabs.diproveboliviapp.ui.theme.BricolageGrotesque
 import com.varqulabs.diproveboliviapp.ui.theme.BricolageGrotesqueSemiCondensed
 
 @OptIn(ExperimentalMaterial3Api::class)
