+++ app/src/main/java/com/varqulabs/diproveboliviapp/divisions/presentation/home/DiproveDivisionsSelector.kt
@@ -1,4 +1,4 @@
-package com.varqulabs.diproveboliviapp.divisions.presentation.components
+package com.varqulabs.diproveboliviapp.divisions.presentation.home
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
