+++ app/src/main/java/com/varqulabs/diproveboliviapp/navigation/Routes.kt
@@ -1,46 +1,49 @@
 package com.varqulabs.diproveboliviapp.navigation
 
 import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
 import com.varqulabs.diproveboliviapp.R
 import kotlinx.serialization.Serializable
 
 @Serializable
-data object WelcomeScreen
-
-@Serializable
-data object HomeScreen
-
-@Serializable
-data object LocationsRoute
-
-@Serializable
-data object ProceduresRoute
-
-@Serializable
-data object DivisionsRoute
-
-@Serializable
-data object DivisionsHomeScreen
-
-@Serializable
-data class DiproveDivisionScreen(val divisionId: Int)
-
-@Serializable
-data object SuggestionsRoute
-
-@Serializable
-sealed class HomeScreens<T>(val name: String, @DrawableRes val icon: Int, val route: T) {
+sealed class Routes(@StringRes val title : Int) {
 
     @Serializable
-    data object DiproveDivisions : HomeScreens<DivisionsRoute>(name = "Divisiones", icon = R.drawable.division_structure_bold, route = DivisionsRoute)
+    data object Welcome : Routes(title = R.string.copy_diprove_cbba)
 
     @Serializable
-    data object Locations : HomeScreens<LocationsRoute>(name = "Direcciones", icon = R.drawable.region_map_rect, route = LocationsRoute)
+    data object Home : Routes(title = R.string.copy_diprove_cbba)
 
     @Serializable
-    data object Suggestions : HomeScreens<SuggestionsRoute>(name = "Recomendaciones", icon = R.drawable.summary_check_rect, route = SuggestionsRoute)
+    data object Locations : Routes(title = R.string.copy_diprove_regional_directorates)
 
     @Serializable
-    data object Procedures : HomeScreens<ProceduresRoute>(name = "Tramites y Servicios", icon = R.drawable.documents_papers, route = ProceduresRoute)
+    data object Procedures : Routes(title = R.string.copy_services_offered)
+
+    @Serializable
+    data object Suggestions : Routes(title = R.string.copy_diprove_suggestions)
+
+    @Serializable
+    data object Divisions_Home : Routes(title = R.string.copy_diprove_cbba_divisions)
+
+    @Serializable
+    data class Divisions_Detail(val divisionId: Int) : Routes(title = R.string.copy_diprove_cbba_divisions)
+
+}
+
+@Serializable
+sealed class HomeScreens<T>(@DrawableRes val icon: Int, val name: String, val route: T) {
+
+    @Serializable
+    data object DiproveDivisions : HomeScreens<Routes>(name = "Divisiones", icon = R.drawable.division_structure_bold, route = Routes.Divisions_Home)
+
+    @Serializable
+    data object Locations : HomeScreens<Routes>(name = "Direcciones", icon = R.drawable.region_map_rect, route = Routes.Locations)
+
+    @Serializable
+    data object Suggestions : HomeScreens<Routes>(name = "Recomendaciones", icon = R.drawable.summary_check_rect, route = Routes.Suggestions)
+
+    @Serializable
+    data object Procedures : HomeScreens<Routes>(name = "Tramites y Servicios", icon = R.drawable.documents_papers, route = Routes.Procedures)
 
 }
