+++ app/src/main/java/com/varqulabs/diproveboliviapp/navigation/utils/NavigateBetweenScreens.kt
@@ -1,0 +1,11 @@
+package com.varqulabs.diproveboliviapp.navigation.utils
+
+import androidx.navigation.NavHostController
+import com.varqulabs.diproveboliviapp.navigation.Routes
+
+fun NavHostController.navigateTo(route: Routes) {
+    navigate(route) {
+        launchSingleTop = true
+        restoreState = true
+    }
+}
