+++ app/src/main/java/com/varqulabs/diproveboliviapp/navigation/utils/SlideInAnimations.kt
@@ -1,0 +1,29 @@
+package com.varqulabs.diproveboliviapp.navigation.utils
+
+import androidx.compose.animation.AnimatedContentTransitionScope
+import androidx.compose.animation.core.tween
+import androidx.navigation.NavBackStackEntry
+
+fun AnimatedContentTransitionScope<NavBackStackEntry>.enterTransition() =
+    slideIntoContainer(
+        towards = AnimatedContentTransitionScope.SlideDirection.Left,
+        animationSpec = tween(600)
+    )
+
+fun AnimatedContentTransitionScope<NavBackStackEntry>.exitTransition() =
+    slideOutOfContainer(
+        towards = AnimatedContentTransitionScope.SlideDirection.Left,
+        animationSpec = tween(600)
+    )
+
+fun AnimatedContentTransitionScope<NavBackStackEntry>.popExitTransition() =
+    slideOutOfContainer(
+        towards = AnimatedContentTransitionScope.SlideDirection.Right,
+        animationSpec = tween(600)
+    )
+
+fun AnimatedContentTransitionScope<NavBackStackEntry>.popEnterTransition() =
+    slideIntoContainer(
+        towards = AnimatedContentTransitionScope.SlideDirection.Right,
+        animationSpec = tween(600)
+    )
