+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Diprove</string>
+    <string name="copy_diprove_cbba">DIPROVE COCHABAMBA</string>
 </resources>
